refactor: move embed template into ai
To not bloat cli and remove embed template together with legacy ai
eventually.
