+++ apps/builder/app/builder/features/ai/ai-fetch-result.ts
@@ -4,6 +4,7 @@ import {
   operations,
   handleAiRequest,
   commandDetect,
+  WsEmbedTemplate,
 } from "@webstudio-is/ai";
 import { createRegularStyleSheet } from "@webstudio-is/css-engine";
 import {
@@ -13,7 +14,6 @@ import {
   componentAttribute,
 } from "@webstudio-is/react-sdk";
 import {
-  type WsEmbedTemplate,
   Instance,
   createScope,
   findTreeInstanceIds,
