+++ apps/builder/package.json
@@ -98,7 +98,7 @@
     "micromark": "^4.0.2",
     "micromark-extension-gfm": "^3.0.0",
     "nanoevents": "^9.1.0",
-    "nanoid": "^5.0.9",
+    "nanoid": "^5.1.5",
     "nanostores": "^0.11.3",
     "picocolors": "^1.1.1",
     "pretty-bytes": "^6.1.1",
