+++ packages/ai/src/utils/jsx-to-template.server.ts
@@ -3,8 +3,8 @@ import {
   jsxToWSEmbedTemplate,
   tailwindToWebstudio,
   traverseTemplate,
+  WsEmbedTemplate,
 } from "@webstudio-is/jsx-utils";
-import { WsEmbedTemplate } from "@webstudio-is/sdk";
 
 export const jsxToTemplate = async (jsx: string) => {
   const template = await jsxToWSEmbedTemplate(jsx);
