+++ packages/asset-uploader/package.json
@@ -18,7 +18,7 @@
     "fontkit": "^2.0.4",
     "image-meta": "^0.2.1",
     "immer": "^10.1.1",
-    "nanoid": "^5.0.9",
+    "nanoid": "^5.1.5",
     "warn-once": "^0.1.1"
   },
   "devDependencies": {
