+++ packages/cli/package.json
@@ -38,7 +38,7 @@
     "change-case": "^5.4.4",
     "deepmerge": "^4.3.1",
     "env-paths": "^3.0.0",
-    "nanoid": "^5.0.9",
+    "nanoid": "^5.1.5",
     "p-limit": "^6.2.0",
     "parse5": "7.2.1",
     "picocolors": "^1.1.1",
