+++ packages/jsx-utils/package.json
@@ -28,10 +28,13 @@
   "sideEffects": false,
   "dependencies": {
     "@webstudio-is/css-data": "workspace:*",
+    "@webstudio-is/css-engine": "workspace:*",
     "@webstudio-is/sdk": "workspace:*",
+    "@webstudio-is/react-sdk": "workspace:*",
     "acorn": "^8.14.1",
     "acorn-jsx": "^5.3.2",
     "json5": "^2.2.3",
+    "nanoid": "^5.1.5",
     "zod": "^3.24.2"
   }
 }
