+++ packages/jsx-utils/src/index.ts
@@ -1,3 +1,4 @@
+export * from "./embed-template";
 export * from "./traverse-template";
 export { jsxToWSEmbedTemplate } from "./jsx";
 export { heroiconsToSvgEmbed } from "./heroicons";
