+++ packages/prisma-client/package.json
@@ -23,7 +23,7 @@
   "license": "AGPL-3.0-or-later",
   "private": true,
   "dependencies": {
-    "nanoid": "^5.0.9",
+    "nanoid": "^5.1.5",
     "tinyexec": "^0.3.2",
     "umzug": "^3.2.1"
   },
