+++ packages/react-sdk/src/index.ts
@@ -1,6 +1,5 @@
 export * from "./remix";
 export * from "./components/components-utils";
-export * from "./embed-template";
 export * from "./props";
 export type * from "./context";
 export type * from "./hook";
