+++ packages/sdk/src/index.ts
@@ -11,7 +11,6 @@ export * from "./schema/styles";
 export * from "./schema/deployment";
 export * from "./schema/webstudio";
 export * from "./schema/prop-meta";
-export * from "./schema/embed-template";
 export * from "./schema/component-meta";
 
 export * from "./core-metas";
